pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "evidence_extractor"
version = "1.0.0"
authors = [
  { name="Aaryan Singh", email="aaryan724singh@gmail.com" },
]
description = "A tool for narrow-domain systematic reviews using LLMs."
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering",
    "Topic :: Text Processing",
]
dependencies = [
    "pydantic>=2.0",
    "click>=8.0",
    "pyyaml>=6.0",
    "pymupdf>=1.23.0",
    "camelot-py[cv]>=0.11.0",
    "opencv-python-headless",
    "thefuzz>=0.20.0",
    "google-generativeai>=0.3.0",
    "python-dotenv>=1.0.0",
    "Pillow>=10.0.0",
    "openpyxl>=3.1.0",
    "graphviz>=0.20.0",
]

[project.urls]
"Homepage" = "https://github.com/ryen-x/eendsr"

[project.scripts]
evidence-extractor = "evidence_extractor.cli.main:cli"

[project.optional-dependencies]
test = [
    "pytest>=7.0",
    "pytest-mock>=3.12.0",
    "ruff>=0.1.0", 
]
notebook = [
    "jupyter>=1.0.0",
]

docs = [
    "sphinx>=7.0.0",
    "furo>=2023.09.10",
]

[tool.pytest.ini_options]
markers = [
    "integration: marks tests as integration tests (slow, may require network)",
]

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "W", "I"] # Standard flake8 rules + isort